Afterpay payment method country available. Rename IPBlockSubscriber.php -> PaymentKeeperSubscriber.php
