+++ custom/plugins/emspay/Resources/config.xml
@@ -27,7 +27,7 @@
             <label>EMS Online Webhook</label>
             <value>true</value>
             <description lang="de">Verwenden Sie den Webhook</description>
-            <description lang="en">Use the cURL CA bundle.</description>
+            <description lang="en">Use the webhook.</description>
             <description lang="nl">Gebruik de webhook</description>
             <description lang="fr">Utilisez le webhook</description>
         </element>
@@ -38,6 +38,14 @@
             <description lang="en">Enter the API Key of the test webshop here for testing Klarna Pay Later. If you do not offer Klarna Pay Later you can leave this empty.</description>
             <description lang="nl">Voer hier de API Key van de test webshop in om Klarna Pay Later te testen. Als u Klarna Pay Later niet aanbiedt, kunt u dit leeg laten.</description>
             <description lang="fr">Saisissez ici la clé API de la boutique en ligne de test pour tester Klarna Pay Later. Si vous ne proposez pas Klarna Pay Later, vous pouvez laisser ce champ vide.</description>
+        </element>
+        <element type="text" scope="locale">
+            <name>emsonline_test_ip_klarna</name>
+            <label>EMS Online Klarna Pay Later Test IP</label>
+            <description lang="de">Sie können bestimmte IP-Adressen angeben, für die Klarna Pay Later sichtbar ist, z. B. wenn Sie Klarna Pay Later testen möchten. Sie können IP-Adressen als 128.0.0.1, 255.255.255.255 eingeben. Wenn Sie dieses Feld leer lassen, ist Klarna Pay Later für alle IP-Adressen sichtbar. "</description>
+            <description lang="en">You can specify specific IP addresses for which the Klarna is used Payment later payment method is visible, eg if you want to test (for example: 128.0.0.1, 255.255.255.255). If you do not enter anything, the payment method is visible for all IP addresses.</description>
+            <description lang="nl">U kunt specifieke IP-adressen specificeren waarvoor Klarna Pay Later zichtbaar is, bijvoorbeeld als u Klarna Pay Later wilt testen. U kunt IP-adressen typen als 128.0.0.1, 255.255.255.255. Als u dit leeg laat, is Klarna Pay Later zichtbaar voor alle IP-adressen. "</description>
+            <description lang="fr">Vous pouvez spécifier des adresses IP spécifiques pour lesquelles Klarna Pay Later est visible, par exemple si vous souhaitez tester Klarna Pay Later. Vous pouvez taper des adresses IP comme 128.0.0.1, 255.255.255.255. Si vous laissez ce Klarna Pay Later vide, il est visible par toutes les adresses IP. "</description>
         </element>
         <element type="text" scope="locale">
             <name>emsonline_test_api_afterpay</name>
@@ -56,12 +64,16 @@
             <description lang="fr">Vous pouvez spécifier des adresses IP spécifiques pour lesquelles AfterPay est visible, par exemple si vous souhaitez tester AfterPay. Vous pouvez taper des adresses IP comme 128.0.0.1, 255.255.255.255. Si vous laissez ce champ AfterPay vide, il est visible par toutes les adresses IP.</description>
         </element>
         <element type="text" scope="locale">
-            <name>emsonline_test_ip_klarna</name>
-            <label>EMS Online Klarna Pay Later Test IP</label>
-            <description lang="de">Sie können bestimmte IP-Adressen angeben, für die Klarna Pay Later sichtbar ist, z. B. wenn Sie Klarna Pay Later testen möchten. Sie können IP-Adressen als 128.0.0.1, 255.255.255.255 eingeben. Wenn Sie dieses Feld leer lassen, ist Klarna Pay Later für alle IP-Adressen sichtbar. "</description>
-            <description lang="en">You can specify specific IP addresses for which the Klarna is used Payment later payment method is visible, eg if you want to test (for example: 128.0.0.1, 255.255.255.255). If you do not enter anything, the payment method is visible for all IP addresses.</description>
-            <description lang="nl">U kunt specifieke IP-adressen specificeren waarvoor Klarna Pay Later zichtbaar is, bijvoorbeeld als u Klarna Pay Later wilt testen. U kunt IP-adressen typen als 128.0.0.1, 255.255.255.255. Als u dit leeg laat, is Klarna Pay Later zichtbaar voor alle IP-adressen. "</description>
-            <description lang="fr">Vous pouvez spécifier des adresses IP spécifiques pour lesquelles Klarna Pay Later est visible, par exemple si vous souhaitez tester Klarna Pay Later. Vous pouvez taper des adresses IP comme 128.0.0.1, 255.255.255.255. Si vous laissez ce Klarna Pay Later vide, il est visible par toutes les adresses IP. "</description>
+            <name>emsonline_afterpay_countries</name>
+            <label>Countries available for AfterPay</label>
+            <label lang="de">Für AfterPay verfügbare Länder</label>
+            <label lang="nl">Landen beschikbaar voor AfterPay</label>
+            <label lang="fr">Pays disponibles pour AfterPay</label>
+            <value>NL, BE</value>
+            <description lang="en">To allow AfterPay to be used for any other country just add its country code (in ISO 2 standard) to the "Countries available for AfterPay" field. Example: BE, NL, FR.</description>
+            <description lang="de">Um AfterPay für jedes andere Land verwenden zu können, fügen Sie einfach seinen Ländercode (in ISO 2-Norm) in das Feld "Für AfterPay verfügbare Länder" ein. Beispiel: BE, NL, FR.</description>
+            <description lang="nl">Om AfterPay voor elk ander land te kunnen gebruiken, voegt u gewoon de landcode (in ISO 2-norm) toe aan het veld "Beschikbare landen voor AfterPay". Voorbeeld: BE, NL, FR.</description>
+            <description lang="fr">Pour autoriser AfterPay à être utilisé pour tout autre pays, ajoutez simplement son code de pays (dans la norme ISO 2) dans le champ "Pays disponibles pour AfterPay". Exemple: BE, NL, FR.</description>
         </element>
 
     </elements>
